ENH: new file 'events' --- starting moving methods that may go into an 'events' object to this file.
